test: remove unneeded plugins
