--- app/tests/helpers/auth_helpers.py
@@ -1,15 +1,0 @@
-from http import cookies
-
-import pytest
-
-
-@pytest.fixture(autouse=False, scope="function")
-def login_operator(client):
-    op_response = client.post(
-        "/auth/login",
-        headers={"accept": "application/json", "Content-Type": "application/json"},
-        json={"username": "test", "password": "test"},
-    )
-    cookie = cookies.SimpleCookie(op_response.headers["set-cookie"])
-    access_token = cookie["access_token"].value
-    return access_token
