--- conftest.py
@@ -1,5 +1,0 @@
-pytest_plugins = [
-    "app.tests.helpers.check_helpers",
-    "app.tests.helpers.report_helpers",
-    "app.tests.helpers.auth_helpers",
-]
